Generate the start script from setup
The particular command I tried using to cd to the start script's directory when starting (needed so lights.py can find config.json in the same directory) wasn't working. Instead of trying to find one magic bash command that does it I decided it'd be easier to just generate the start script like it's already doing with the service, so I can use pwd to embed the right path, so as long as setup is run from the right directory things should work fine.